Refactor API endpoints: Updated data fetching hooks to remove tenant-specific paths, enhancing code simplicity and maintainability. Deleted obsolete route files related to tenant-specific data handling.
